+++ src/app/models/overallInfo.model.ts
@@ -1,8 +1,10 @@
-export interface OverallLeft {
-  overall: number;
-  left: number;
+import {Money} from "./settings.model";
+
+export interface OverallLeft<Type> {
+  overall: Type;
+  left: Type;
 }
 
-export type DaysInfo = OverallLeft;
-export type MoneyInfo = OverallLeft;
-export type BudgetInfo = OverallLeft;
+export type DaysInfo = OverallLeft<number>;
+export type MoneyInfo = OverallLeft<Money>;
+export type BudgetInfo = OverallLeft<Money>;