+++ src/app/state/publicSelectors/dayBudget.selector.ts
@@ -3,10 +3,17 @@ import {selectBudgetType} from "../selectors/settings.selector";
 import {BudgetType} from "../../models/settings.model";
 import {BudgetInfo} from "../../models/overallInfo.model";
 import {selectStaticDayBudget} from "../selectors/dayBudget.static.selector";
+import {selectDynamicDayBudget} from "../selectors/dayBudget.dynamic.selector";
 
 export const selectDayBudget = createSelector(
   selectBudgetType,
   selectStaticDayBudget,
-  (budgetType: BudgetType, staticDB): BudgetInfo =>
-    staticDB
+  selectDynamicDayBudget,
+  (
+    budgetType: BudgetType,
+    staticDB: BudgetInfo,
+    dynamicDB: BudgetInfo
+  ): BudgetInfo =>
+    budgetType === BudgetType.Dynamic ?
+      dynamicDB : staticDB
 )