+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Budget Type: Static or Dynamic 
 - Static Budget Type implementation
+- Dynamic Budget Type implementation
 ### Fix
 - show savings as part of existing money